api/urls.py
from django.urls import path , include 
from . import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)


urlpatterns=[
    path('users/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('users/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    path('' , views.getRoutes ) ,
    path('projects/' , views.getProjects),
    path('project/<uuid:pk>' , views.getProject),
    path('vote/<uuid:pk>' , views.projectVote),

]